back/requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cors==1.0.1
cssselect==1.2.0
exceptiongroup==1.1.0
fake-useragent==1.1.3
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
isort==5.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
json-normalize==1.0.1
jsonify==0.5
lazy-object-proxy==1.4.3
load-dotenv==0.1.0
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.6.1
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==2.0.0
parse==1.19.0
pyee==8.2.2
pylint==2.6.0
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
selenium==4.8.2
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.3.18
tldextract==3.4.0
toml==0.10.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.12.1
wsproto==1.2.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
